Cache reads of metadata in RAR
Improved implementation of 55f846e7: doesn't hold handles to the
assemblies after reading their metadata.

Long term, more functionality could be moved into AssemblyMetadata and
out of AssemblyInformation, and `GetAssemblyMetadata()` could be modified
to return an `AssemblyMetadata` instead of its three out parameters. But
that's a risky change to make at this point in the cycle.